Snapshot of 19 December 2025
This code compiles but problem areas like the 'foreach' constructs are commented out it will not link. The DEFINE_FLEX is converted into written out code because it needs the TypeOf which is missed by the ROOL DDE. In al honesty in my idea you have a better idea what happens. WRITE_ONCE has been implemented with a simple macro but in reality some it should be an atomic operation. At this moment it is less important because RO is not multithreaded / or multiprocessor supported ;-).